+++ src/main/java/com/mmz/specs/service/MaterialListService.java
@@ -34,10 +34,9 @@ public interface MaterialListService {
     public void removeMaterialList(int id);
 
 
-    public MaterialListEntity getMaterialListById(int id);
+    public MaterialListEntity getMaterialListById(long id);
 
     public List<MaterialListEntity> getMaterialListByDetail(DetailEntity detailEntity);
-
 
 
     public List<MaterialListEntity> listMaterialLists();
